requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiohttp==3.10.0
psycopg2-binary>=2.9
